requirements.txt
torch>=2.2.0
transformers>=4.41.0
datasets>=2.19.0
scikit-learn>=1.4.0
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0
numpy>=1.26.0
pandas>=2.2.0
tqdm>=4.66.0
streamlit>=1.35.0
python-dotenv>=1.0.1
